+++ test/framework/src/test/java/org/elasticsearch/test/test/ESTestCaseTests.java
@@ -23,7 +23,6 @@ import junit.framework.AssertionFailedError;
 
 import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.xcontent.XContentBuilder;
-import org.elasticsearch.common.xcontent.XContentFactory;
 import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.common.xcontent.XContentType;
 import org.elasticsearch.test.ESTestCase;
@@ -39,9 +38,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Predicate;
 
-import static org.elasticsearch.test.XContentTestUtils.insertRandomFields;
 import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.hasSize;
 
@@ -166,81 +163,5 @@ public class ESTestCaseTests extends ESTestCase {
 
     public void testRandomUniqueNormalUsageAlwayMoreThanOne() {
         assertThat(randomUnique(() -> randomAlphaOfLengthBetween(1, 20), 10), hasSize(greaterThan(0)));
-    }
-
-    @SuppressWarnings("unchecked")
-    public void testInsertRandomXContent() throws IOException {
-        XContentBuilder builder = XContentFactory.jsonBuilder();
-        builder.startObject();
-        {
-            builder.startObject("foo");
-            {
-                builder.field("bar", 1);
-            }
-            builder.endObject();
-            builder.startObject("foo1");
-            {
-                builder.startObject("foo2");
-                {
-                    builder.field("buzz", 1);
-                }
-                builder.endObject();
-            }
-            builder.endObject();
-            builder.field("foo3", 2);
-            builder.startArray("foo4");
-            {
-                builder.startObject();
-                {
-                    builder.field("foo5", 1);
-                }
-                builder.endObject();
-            }
-            builder.endArray();
-        }
-        builder.endObject();
-
-        Map<String, Object> resultMap;
-
-        try (XContentParser parser = createParser(XContentType.JSON.xContent(),
-                insertRandomFields(builder.contentType(), builder.bytes(), null, random()))) {
-            resultMap = parser.map();
-        }
-        assertEquals(5, resultMap.keySet().size());
-        assertEquals(2, ((Map<String, Object>) resultMap.get("foo")).keySet().size());
-        Map<String, Object> foo1 = (Map<String, Object>) resultMap.get("foo1");
-        assertEquals(2, foo1.keySet().size());
-        assertEquals(2, ((Map<String, Object>) foo1.get("foo2")).keySet().size());
-        List<Object> foo4List = (List<Object>) resultMap.get("foo4");
-        assertEquals(1, foo4List.size());
-        assertEquals(2, ((Map<String, Object>) foo4List.get(0)).keySet().size());
-
-        Predicate<String> pathsToExclude = path -> path.endsWith("foo1");
-        try (XContentParser parser = createParser(XContentType.JSON.xContent(),
-                insertRandomFields(builder.contentType(), builder.bytes(), pathsToExclude, random()))) {
-            resultMap = parser.map();
-        }
-        assertEquals(5, resultMap.keySet().size());
-        assertEquals(2, ((Map<String, Object>) resultMap.get("foo")).keySet().size());
-        foo1 = (Map<String, Object>) resultMap.get("foo1");
-        assertEquals(1, foo1.keySet().size());
-        assertEquals(2, ((Map<String, Object>) foo1.get("foo2")).keySet().size());
-        foo4List = (List<Object>) resultMap.get("foo4");
-        assertEquals(1, foo4List.size());
-        assertEquals(2, ((Map<String, Object>) foo4List.get(0)).keySet().size());
-
-        pathsToExclude = path -> path.contains("foo1");
-        try (XContentParser parser = createParser(XContentType.JSON.xContent(),
-                insertRandomFields(builder.contentType(), builder.bytes(), pathsToExclude, random()))) {
-            resultMap = parser.map();
-        }
-        assertEquals(5, resultMap.keySet().size());
-        assertEquals(2, ((Map<String, Object>) resultMap.get("foo")).keySet().size());
-        foo1 = (Map<String, Object>) resultMap.get("foo1");
-        assertEquals(1, foo1.keySet().size());
-        assertEquals(1, ((Map<String, Object>) foo1.get("foo2")).keySet().size());
-        foo4List = (List<Object>) resultMap.get("foo4");
-        assertEquals(1, foo4List.size());
-        assertEquals(2, ((Map<String, Object>) foo4List.get(0)).keySet().size());
     }
 }
